relocating dashboard tab js code to its own file js linting
